Replace the version command with a more generic dev info command
