+++ modules/regexes.js
@@ -1,0 +1,6 @@
+/**
+ * This module contains useful regular expressions
+ * @module regexes
+ */
+
+module.exports.numeric = /^\d+$/;
